+++ lib/src/maintenance.dart
@@ -14,7 +14,6 @@ import 'package:yaml/yaml.dart' as yaml;
 
 import 'dartdoc_analyzer.dart';
 import 'download_utils.dart';
-import 'messages.dart';
 import 'model.dart';
 import 'pubspec.dart';
 import 'utils.dart';
@@ -65,16 +64,19 @@ String firstFileFromNames(List<String> files, List<String> names,
 }
 
 double getMaintenanceScore(Maintenance maintenance, {Duration age}) {
-  return applyPenalties(
-      1.0, _getAllSuggestion(maintenance, age: age)?.map((s) => s.penalty));
-}
-
-List<Suggestion> _getAllSuggestion(Maintenance maintenance, {Duration age}) {
-  final list = <Suggestion>[];
+  var score = 100.0;
+  maintenance?.suggestions
+      ?.map((s) => s.score)
+      ?.where((d) => d != null)
+      ?.forEach((d) {
+    score -= d;
+  });
   final ageSuggestion = getAgeSuggestion(age);
-  if (ageSuggestion != null) list.add(ageSuggestion);
-  if (maintenance.suggestions != null) list.addAll(maintenance.suggestions);
-  return list;
+  if (ageSuggestion?.score != null) {
+    score -= ageSuggestion.score;
+  }
+
+  return max(0.0, min(100.0, score));
 }
 
 Suggestion getAgeSuggestion(Duration age) {
@@ -85,19 +87,19 @@ Suggestion getAgeSuggestion(Duration age) {
         SuggestionCode.packageVersionObsolete,
         'Package is too old.',
         'The package was released more than two years ago.',
-        penalty: new Penalty(amount: 10000));
+        score: 100.0);
   }
 
   // adjust score to the age
   if (age > _year) {
     final ageInWeeks = age.inDays ~/ 7;
     final daysOverAYear = age.inDays - _year.inDays;
-    final basisPoints = daysOverAYear * 10000 ~/ 365;
+    final score = max(0.0, min(100.0, daysOverAYear * 100.0 / 365));
     return new Suggestion.hint(
         SuggestionCode.packageVersionOld,
         'Package is getting outdated.',
         'The package was released ${ageInWeeks} weeks ago.',
-        penalty: new Penalty(fraction: min(10000, max(0, basisPoints))));
+        score: score);
   }
 
   return null;
@@ -107,8 +109,6 @@ Future<Maintenance> detectMaintenance(
   UrlChecker urlChecker,
   String pkgDir,
   Pubspec pubspec,
-  List<CodeProblem> analyzerItems,
-  List<Suggestion> dartFileSuggestions,
   List<PkgDependency> unconstrainedDeps, {
   @required DartPlatform pkgPlatform,
   bool dartdocSuccessful,
@@ -173,14 +173,14 @@ Future<Maintenance> detectMaintenance(
       SuggestionCode.pubspecHomepageIsNotHelpful,
       'Homepage is not helpful.',
       'Update the `homepage` property: create a website about the package or use the source repository URL.',
-      penalty: new Penalty(fraction: 1000),
+      score: 10.0,
     ));
   } else if (homepageStatus == UrlStatus.missing) {
     maintenanceSuggestions.add(new Suggestion.warning(
       SuggestionCode.pubspecHomepageDoesNotExists,
       'Homepage does not exists.',
       'We were unable to access `${pubspec.homepage}` at the time of the analysis.',
-      penalty: new Penalty(fraction: 1000),
+      score: 20.0,
     ));
   }
 
@@ -192,14 +192,14 @@ Future<Maintenance> detectMaintenance(
         SuggestionCode.pubspecDocumentationIsNotHelpful,
         'Documentation URL is not helpful.',
         'Update the `documentation` property: create a website about the package or remove it.',
-        penalty: new Penalty(fraction: 100),
+        score: 10.0,
       ));
     } else if (documentationStatus == UrlStatus.missing) {
       maintenanceSuggestions.add(new Suggestion.warning(
         SuggestionCode.pubspecDocumentationDoesNotExists,
         'Documentation URL does not exists.',
         'We were unable to access `${pubspec.documentation}` at the time of the analysis.',
-        penalty: new Penalty(fraction: 500),
+        score: 10.0,
       ));
     }
   }
@@ -210,7 +210,7 @@ Future<Maintenance> detectMaintenance(
         'Add SDK constraint in `pubspec.yaml`.',
         'For information about setting SDK constraint, please see '
         '[https://www.dartlang.org/tools/pub/pubspec#sdk-constraints](https://www.dartlang.org/tools/pub/pubspec#sdk-constraints).',
-        penalty: new Penalty(fraction: 500)));
+        score: 5.0));
   }
 
   if (pubspec.shouldWarnDart2Constraint) {
@@ -220,7 +220,7 @@ Future<Maintenance> detectMaintenance(
         'The SDK constraint in pubspec.yaml doesn\'t allow the Dart 2.0.0 release. '
         'For information about upgrading it to be Dart 2 compatible, please see '
         '[https://www.dartlang.org/dart-2#migration](https://www.dartlang.org/dart-2#migration).',
-        penalty: new Penalty(fraction: 2000)));
+        score: 20.0));
   }
 
   if (dartdocSuccessful == false) {
@@ -232,7 +232,7 @@ Future<Maintenance> detectMaintenance(
         SuggestionCode.platformConflictInPkg,
         'Fix platform conflicts.',
         pkgPlatform.reason,
-        penalty: new Penalty(fraction: 2000)));
+        score: 20.0));
   }
 
   if (!changelogExists) {
@@ -240,14 +240,14 @@ Future<Maintenance> detectMaintenance(
         SuggestionCode.changelogMissing,
         'Maintain `CHANGELOG.md`.',
         'Changelog entries help clients to follow the progress in your code.',
-        penalty: new Penalty(fraction: 2000)));
+        score: 20.0));
   }
   if (!readmeExists) {
     maintenanceSuggestions.add(new Suggestion.warning(
         SuggestionCode.readmeMissing,
         'Maintain `README.md`.',
         'Readme should inform others about your project, what it does, and how they can use it.',
-        penalty: new Penalty(fraction: 500)));
+        score: 30.0));
   }
   if (!exampleExists) {
     final exampleDirExists = files.any((file) => file.startsWith('example/'));
@@ -256,29 +256,30 @@ Future<Maintenance> detectMaintenance(
           SuggestionCode.exampleMissing,
           'Maintain an example.',
           'None of the files in your `example/` directory matches a known example patterns. '
-          'Common file name patterns include: `main.dart`, `example.dart` or you could also use `$pkgName.dart`.',
-          penalty: new Penalty(amount: 1)));
+          'Common file name patterns include: `main.dart`, `example.dart` or you could also use `$pkgName.dart`.'));
     } else {
       maintenanceSuggestions.add(new Suggestion.hint(
           SuggestionCode.exampleMissing,
           'Maintain an example.',
           'Create a short demo in the `example/` directory to show how to use this package. '
           'Common file name patterns include: `main.dart`, `example.dart` or you could also use `$pkgName.dart`.',
-          penalty: new Penalty(amount: 5)));
+          score: 10.0));
     }
   }
   if (oldAnalysisOptions) {
-    maintenanceSuggestions.add(new Suggestion.hint(
+    maintenanceSuggestions.add(new Suggestion.warning(
         SuggestionCode.analysisOptionsRenameRequired,
         'Use `analysis_options.yaml`.',
-        'Rename old `.analysis_options` file to `analysis_options.yaml`.'));
+        'Rename old `.analysis_options` file to `analysis_options.yaml`.',
+        score: 10.0));
   }
   if (analysisOptionsExists && strongModeDisabled) {
-    maintenanceSuggestions.add(new Suggestion.hint(
+    maintenanceSuggestions.add(new Suggestion.warning(
         SuggestionCode.analysisOptionsWeakMode,
         'The option `strong-mode: false` is being deprecated.',
         'Remove `strong-mode: false` from your `analysis_options.yaml` file:\n\n'
-        '```\nanalyzer:\n  strong-mode: false\n```\n'));
+        '```\nanalyzer:\n  strong-mode: false\n```\n',
+        score: 50.0));
   }
 
   final version = pubspec.version;
@@ -293,7 +294,7 @@ Future<Maintenance> detectMaintenance(
         'While there is nothing inherently wrong with versions of `0.*.*`, it '
         'usually means that the author is still experimenting with the general '
         'direction of the API.',
-        penalty: new Penalty(amount: 10)));
+        score: 10.0));
   }
 
   // Not a "gold" release
@@ -303,7 +304,7 @@ Future<Maintenance> detectMaintenance(
         'Package is pre-release.',
         'Pre-release versions should be used with caution, their API may change '
         'in breaking ways.',
-        penalty: new Penalty(fraction: 200)));
+        score: 5.0));
   }
 
   // Checking the length of description.
@@ -315,82 +316,21 @@ Future<Maintenance> detectMaintenance(
         'Description is critical to giving users a quick insight into the features '
         'of the package and why it is relevant to their query. '
         'Ideal length is between 60 and 180 characters.',
-        penalty: new Penalty(fraction: 500)));
+        score: 20.0));
   } else if (description.length < 60) {
     maintenanceSuggestions.add(new Suggestion.hint(
         SuggestionCode.pubspecDescriptionTooShort,
         'The description is too short.',
         'Add more detail about the package, what it does and what is its target use case. '
         'Try to write at least 60 characters.',
-        penalty: new Penalty(amount: 20)));
+        score: 20.0));
   } else if (description.length > 180) {
     maintenanceSuggestions.add(new Suggestion.hint(
         SuggestionCode.pubspecDescriptionTooLong,
         'The description is too long.',
         'Search engines will display only the first part of the description. '
         'Try to keep it under 180 characters.',
-        penalty: new Penalty(amount: 10)));
-  }
-
-  final errorCount = analyzerItems.where((s) => s.isError).length;
-  final warningCount = analyzerItems.where((s) => s.isWarning).length;
-  final hintCount = analyzerItems.where((s) => s.isInfo).length;
-
-  final reportedFiles = new Set();
-  final onePerFileSuggestions = dartFileSuggestions
-      .where((s) => reportedFiles.add(s.file))
-      .toList()
-        ..sort();
-
-  if (onePerFileSuggestions.length < 5) {
-    maintenanceSuggestions.addAll(onePerFileSuggestions);
-  } else {
-    final topSuggestions = onePerFileSuggestions.take(2).toList();
-    final restSuggestions = onePerFileSuggestions.skip(2).toList();
-    maintenanceSuggestions.addAll(topSuggestions);
-
-    if (restSuggestions.isNotEmpty) {
-      final bulkErrorCount = restSuggestions.where((s) => s.isError).length;
-      final bulkWarningCount = restSuggestions.where((s) => s.isWarning).length;
-      final bulkHintCount = restSuggestions.where((s) => s.isHint).length;
-      final sb = new StringBuffer();
-      sb.write('Additional issues in the following files:\n\n');
-
-      for (var s in restSuggestions) {
-        final fileAnalyzerItems =
-            analyzerItems.where((cp) => cp.file == s.file).toList();
-
-        if (fileAnalyzerItems.isNotEmpty) {
-          final errorCount = fileAnalyzerItems.where((cp) => cp.isError).length;
-          final warningCount =
-              fileAnalyzerItems.where((cp) => cp.isWarning).length;
-          final hintCount = fileAnalyzerItems.where((cp) => cp.isInfo).length;
-          final issueCounts =
-              formatIssueCounts(errorCount, warningCount, hintCount);
-          sb.writeln('- `${s.file}` ($issueCounts)');
-        } else {
-          sb.writeln('- `${s.file}` (${s.description})');
-        }
-      }
-
-      final level = (bulkErrorCount > 0 || bulkWarningCount > 0)
-          ? SuggestionLevel.warning
-          : SuggestionLevel.hint;
-      maintenanceSuggestions.add(
-        new Suggestion(
-          SuggestionCode.bulk,
-          level,
-          'Fix additional ${restSuggestions.length} files with analysis or formatting issues.',
-          sb.toString(),
-          // These are already reflected in the fitness score, but we'll also
-          // penalize them here (with a much smaller amount), reflecting the need
-          // of work.
-          penalty: new Penalty(
-              amount:
-                  bulkErrorCount * 50 + bulkWarningCount * 10 + bulkHintCount),
-        ),
-      );
-    }
+        score: 10.0));
   }
 
   if (unconstrainedDeps != null && unconstrainedDeps.isNotEmpty) {
@@ -405,7 +345,7 @@ Future<Maintenance> detectMaintenance(
         'Use constrained dependencies.',
         'The `pubspec.yaml` contains $pluralized without version constraints. '
         'Specify version ranges for the following dependencies: $names.',
-        penalty: new Penalty(fraction: 500)));
+        score: 20.0));
   }
 
   maintenanceSuggestions.sort();
@@ -419,19 +359,6 @@ Future<Maintenance> detectMaintenance(
     isExperimentalVersion: isExperimentalVersion,
     isPreReleaseVersion: isPreReleaseVersion,
     dartdocSuccessful: dartdocSuccessful,
-    errorCount: errorCount,
-    warningCount: warningCount,
-    hintCount: hintCount,
     suggestions: maintenanceSuggestions,
   );
-}
-
-double applyPenalties(double initialScore, Iterable<Penalty> penalties) {
-  if (penalties == null) return initialScore;
-  var score = initialScore;
-  for (var p in penalties) {
-    if (p == null) continue;
-    score = p.apply(score);
-  }
-  return score;
 }