+++ lib/src/version.g.dart
@@ -6,4 +6,4 @@ part of pana.version;
 // PackageVersionGenerator
 // **************************************************************************
 
-final _$panaPkgVersionPubSemverVersion = new Version.parse("0.11.9");
+final _$panaPkgVersionPubSemverVersion = new Version.parse("0.12.0-dev");