+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.12.0-dev
+
+**Breaking changes:**
+
+* Named parameters in the following constructors: `Summary`, `DartFileSummary`, `PkgDependency`, `CodeProblem`. 
+
+* Removed per-file and overall `Fitness`, using a top-level `Health` report instead.
+  Follows the proposed changes that allow us to clearly communicate the scoring mechanism.
+
+* Removed `Penalty` and using the simpler to understand `score` in place of it.
+
 ## 0.11.8
 
 * Support Dart 2 gold release.